Move `visual_line_count` and associated test to draw_target.rs
